// src/app/shared/data-access/auth/interfaces/auth-actions.interface.ts
export enum AuthActions {
  LOGIN = '[Auth] Login',
  REGISTER = '[Auth] Register',
  RESET_PASSWORD = '[Auth] Reset password',
  LOGIN_SUCCESS = '[Auth] Login success',
  LOGIN_FAILURE = '[Auth] Login failure',
  REGISTER_SUCCESS = '[Auth] Register success',
  REGISTER_FAILURE = '[Auth] Register failure',
  USER_INFO = '[Auth] User info',
  USER_INFO_SUCCESS = '[Auth] User info success',
  USER_INFO_FAILURE = '[Auth] User info failure',
  LOGOUT = '[Auth] Logout',
  LOGOUT_SUCCESS = '[Auth] Logout success',
  LOGOUT_FAILURE = '[Auth] Logout failure',
}